requirements.txt
tensorflow==2.4.1
pillow
wavio
kapre
pandas
librosa
